Move ViewModel to Hydrogen.Core namespace
