+++ src/Hydrogen/Hydrogen/ViewLocator.cs
@@ -1,7 +1,7 @@
 using System;
 using Avalonia.Controls;
 using Avalonia.Controls.Templates;
-using Hydrogen.ViewModels;
+using Hydrogen.Core;
 
 namespace Hydrogen;
