+++ src/Hydrogen/Hydrogen/ViewModels/MainViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using CommunityToolkit.Mvvm.ComponentModel;
+using Hydrogen.Core;
 using Hydrogen.Models;
 
 namespace Hydrogen.ViewModels;
