+++ src/Hydrogen/Hydrogen/ViewModels/MenuViewModel.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using CommunityToolkit.Mvvm.ComponentModel;
+using Hydrogen.Core;
 using Hydrogen.Models;
 
 namespace Hydrogen.ViewModels;
